+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_late_keyword/avoid_late_keyword_rule_test.dart
@@ -29,7 +29,7 @@ void main() {
         startColumns: [3, 3, 5, 5, 1, 1],
         locationTexts: [
           "late final field = 'string'",
-          'late String uninitializedField',
+          'late int uninitializedField',
           "late final variable = 'string'",
           'late String uninitializedVariable',
           "late final topLevelVariable = 'string'",
@@ -41,6 +41,47 @@ void main() {
           "Avoid using 'late' keyword.",
           "Avoid using 'late' keyword.",
           "Avoid using 'late' keyword.",
+          "Avoid using 'late' keyword.",
+        ],
+      );
+    });
+
+    test('reports about found issues with allow initialized config', () async {
+      final unit = await RuleTestHelper.resolveFromFile(_examplePath);
+      final issues =
+          AvoidLateKeywordRule({'allow-initialized': true}).check(unit);
+
+      RuleTestHelper.verifyIssues(
+        issues: issues,
+        startLines: [8, 17, 23],
+        startColumns: [3, 5, 1],
+        locationTexts: [
+          'late int uninitializedField',
+          'late String uninitializedVariable',
+          'late String topLevelUninitializedVariable',
+        ],
+        messages: [
+          "Avoid using 'late' keyword.",
+          "Avoid using 'late' keyword.",
+          "Avoid using 'late' keyword.",
+        ],
+      );
+    });
+
+    test('reports about found issues with ignored types config', () async {
+      final unit = await RuleTestHelper.resolveFromFile(_examplePath);
+      final issues = AvoidLateKeywordRule({
+        'ignored-types': ['String'],
+      }).check(unit);
+
+      RuleTestHelper.verifyIssues(
+        issues: issues,
+        startLines: [8],
+        startColumns: [3],
+        locationTexts: [
+          'late int uninitializedField',
+        ],
+        messages: [
           "Avoid using 'late' keyword.",
         ],
       );