+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_moving_to_variable/examples/prefix_example.dart
@@ -1,0 +1,19 @@
+import 'example.dart' as prefix;
+import 'generics_example.dart';
+
+void main() {
+  AnotherEnum.anotherValue;
+  AnotherEnum.anotherValue;
+  AnotherEnum.firstValue;
+
+  prefix.SomeValue.firstValue;
+  prefix.SomeValue.firstValue;
+  prefix.SomeValue.secondValue;
+
+  prefix.SomeClass.value;
+  prefix.SomeClass.value;
+  prefix.instance.field;
+
+  print(prefix.SomeValue.entry1);
+  print(prefix.SomeValue.entry2);
+}