+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_moving_to_variable/prefer_moving_to_variable_rule_test.dart
@@ -15,6 +15,10 @@ const _cascadeExamplePath =
     'prefer_moving_to_variable/examples/cascade_example.dart';
 const _genericsExamplePath =
     'prefer_moving_to_variable/examples/generics_example.dart';
+const _assignmentExamplePath =
+    'prefer_moving_to_variable/examples/assignment_example.dart';
+const _prefixExamplePath =
+    'prefer_moving_to_variable/examples/prefix_example.dart';
 
 void main() {
   group('PreferMovingToVariableRule', () {
@@ -215,8 +219,22 @@ void main() {
       RuleTestHelper.verifyNoIssues(issues);
     });
 
+    test('reports no issues for assignments', () async {
+      final unit = await RuleTestHelper.resolveFromFile(_assignmentExamplePath);
+      final issues = PreferMovingToVariableRule().check(unit);
+
+      RuleTestHelper.verifyNoIssues(issues);
+    });
+
     test('reports no issues for generics', () async {
       final unit = await RuleTestHelper.resolveFromFile(_genericsExamplePath);
+      final issues = PreferMovingToVariableRule().check(unit);
+
+      RuleTestHelper.verifyNoIssues(issues);
+    });
+
+    test('reports no issues for prefix imports', () async {
+      final unit = await RuleTestHelper.resolveFromFile(_prefixExamplePath);
       final issues = PreferMovingToVariableRule().check(unit);
 
       RuleTestHelper.verifyNoIssues(issues);